bug fix: see the rest of this comment in acceptor: We need to watch to see if the submitter dies. If it does, there is a chance that we might be the only remaining record of this txn and so we need to ensure progress somehow. To see how this happens, consider the following scenario:
1. Provided the submitter stays up, its repeating sender will
make sure that the txn gets to all proposers, and progress
continues to be made.

2. But consider what happens if the submitter and a proposer are
on the same node which fails: That proposer has local votes and
has sent those votes to us, so we now contain state. But that
node now goes now. The txn never made it to any other node (we
must be an acceptor, and a learner), so when the node comes back
up, there is no record of it anywhere, other than in any such
acceptor.

Once we've gone to disk, we will then have a repeating 2B sender
which will ensure progress, so we have no risk once we've
started going to disk.

However, if we are a learner, then we cannot start an abort
proposer as we're not allowed to vote. So our response in this
scenario is actually to start a repeating sender of the txn
itself to the other active RMs, thus taking the role of the
submitter.

The other thing this issue has revealed is that the proposer cannot send deflated txns because we might need the full txn as an acceptor to send on in the above scenario. So we can only deflate after we have a result in the acceptor.

--HG--
branch : dev